+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - ...
 - feat: Add optional TypeScript support (BYOTS) – (Bring Your Own TypeScript)
 - feat: Use the React plugin's native `version: "detect"` feature
+- fix: Only use `yarn` for installing if a `yarn.lock` file is found
 
 ## 0.2.7